Attempt refactor to allow namesapces but failed: scripts with no class not finding classes in namesapces
