+++ App/Controllers/sitepages.controller.php
@@ -5,6 +5,5 @@ $pdostmt_sitepages = DBManager::query(
     $columns="uri"
 );
 
-# echo "<hr><h1>Site Pages Controller Loaded</h1>";
 require "./Public/Views/sitepages.view.php";
 ?>
