+++ Core/Configuration.php
@@ -1,12 +1,16 @@
 <?php
 
-App::set(
-    "default_database",
-    [
-        "dsn" => "mysql:host=db-wtperfmon.chvhk2xcrgnq.ap-southeast-1.rds.amazonaws.com;dbname=wt_perfmon;",
-        "username" => "admin",
-        "password" => "vVYb9mvF",
-        "options" => []
-    ]
-);
+// namespace Core;
+
+// use App\AppSpecific\App;
+
+// App::set(
+//     $alias="default_database",
+//     $dependency=[
+//         "dsn" => "mysql:host=db-wtperfmon.chvhk2xcrgnq.ap-southeast-1.rds.amazonaws.com;dbname=wt_perfmon;",
+//         "username" => "admin",
+//         "password" => "vVYb9mvF",
+//         "options" => []
+//     ]
+// );
 ?>
