+++ Public/Views/about.view.php
@@ -1,5 +1,5 @@
 <h1>About</h1>
 
 <?php
-include "snippets/footer.php";
+include_once "./Public/snippets/footer.php";
 ?>
