+++ index.php
@@ -1,12 +1,17 @@
 <?php
+namespace Core\Routing;
+
 // Options to show warnings/errors.
 error_reporting(E_ALL);
 ini_set('display_errors', '1');
 
-// require "./vendor/autoload.php";
+//Required scripts.
 require_once "./vendor/autoload.php";
 require_once "./Core/Bootstrap.php";
+require_once "./Core/Routing/Router.php";
+require_once "./Core/Routing/RouteRegistry.php";
 
+// Standard non-production code for monitoring and testing purposes.
 echo "<hr>RAW URI:<br>";
 var_dump($_SERVER["REQUEST_URI"]);
 
@@ -17,9 +22,8 @@ var_dump($passed_uri);
 echo "<hr>Request Type:<br>";
 var_dump(Router::getReqeustType());
 
-$filepath = Router::returnPath($passed_uri, Router::getReqeustType());
-echo "<hr>Fetched Filepath:<br>$filepath";
-
-require $filepath;
-
+Router::callController(
+    $uri=$passed_uri,
+    $request_type=$_SERVER["REQUEST_URI"]
+)
 ?>
